+++ types/codec.go
@@ -1,0 +1,7 @@
+package types
+
+import "github.com/ginchuco/ginchu/codec"
+
+var (
+	cdc = codec.Protobuf{}
+)